+++ Console/CommandLibrary/StatCommand.cs
@@ -19,5 +19,26 @@ namespace ConsoleUtility
         {
             Console.RegisterView<StatView>();
         }
+
+        class StatView : View
+        {
+            public StatView() : base(15f) { }
+
+            public override string GetDebugViewString()
+            {
+                return $@"
+    Current Scene: {UnityEngine.SceneManagement.SceneManager.GetActiveScene().name}
+
+    Time Since Level Load : {Time.timeSinceLevelLoad} seconds
+    FPS : {1f / Time.deltaTime}
+    Delta Time : {Time.smoothDeltaTime * 1000}ms 
+    Time Scale : {Time.timeScale}x
+
+    Camera : {Camera.main?.name}
+    Position : {Camera.main?.transform.position}
+
+";
+            }
+        }
     }
 }
