--- Console/DebugViews.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3b07bd2ff5bec984892f382bbf1a59eb
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
