--- Console/DebugViews/StatView.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ff0bdf8dd3f61e84f89d845a2679911e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
